Add rank slider bar for navigating siblings
Usecase: nodes that have e.g. 20000 children.

With that GUI element it is sufficient to select
any of the child nodes and jump to the n-th sibling
via moving the slider or entering n into the edit
field.

The rank of a node is also displayed as the node's
tool-tip.

The rank edit-field can quickly be accessed
via a shortcut (Ctrl+G -> goto), as well.